improve debt tracking UX with dialog-based add/payment flow

+++ app/src/main/java/com/example/debttracker/data/local/dao/DebtDao.kt
@@ -18,5 +18,5 @@ interface DebtDao {
     suspend fun insertDebt(debt: Debt): Long
 
     @Query("UPDATE debts SET currentAmount = :amount WHERE id = :id")
-    suspend fun updateCurrentAmount(id: Long, amount: Double)
+    suspend fun updateCurrentAmount(id: Long, amount: Long)
 }
